@@ -1,5 +1,5 @@
-module github.com/containerd/btrfs
+module github.com/containerd/btrfs/v2
 
-go 1.15
+go 1.19
 
-require github.com/pkg/errors v0.9.1
+require golang.org/x/sys v0.5.0
