@@ -14,23 +14,25 @@
    limitations under the License.
 */
 
-#include <stddef.h>
-#include <linux/magic.h>
-#include <btrfs/ioctl.h>
-#include <btrfs/ctree.h>
+#include <linux/version.h>
+#if LINUX_VERSION_CODE < KERNEL_VERSION(4,12,0)
+#error "Headers from kernel >= 4.12 are required on compilation time (not on run time)"
+#endif
+#include <linux/btrfs.h>
+#include <linux/btrfs_tree.h>
 
 // unfortunately, we need to define "alignment safe" C structs to populate for
 // packed structs that aren't handled by cgo. Fields will be added here, as
 // needed.
 
 struct gosafe_btrfs_root_item {
-	u8 uuid[BTRFS_UUID_SIZE];
-	u8 parent_uuid[BTRFS_UUID_SIZE];
-	u8 received_uuid[BTRFS_UUID_SIZE];
+	__u8 uuid[BTRFS_UUID_SIZE];
+	__u8 parent_uuid[BTRFS_UUID_SIZE];
+	__u8 received_uuid[BTRFS_UUID_SIZE];
 
-	u64 gen;
-	u64 ogen;
-	u64 flags;
+	__le64 generation;
+	__le64 otransid;
+	__le64 flags;
 };
 
 void unpack_root_item(struct gosafe_btrfs_root_item* dst, struct btrfs_root_item* src);
